[SVLS-9359] move env trackers to internal module and harden git tag upsert
