+++ src/env.ts
@@ -8,6 +8,7 @@
 
 import log from "loglevel";
 import { runtimeLookup, RuntimeType } from "./constants";
+import { setTrackedEnv, getTrackedEnv, hasTrackedEnv } from "./env-tracker";
 import { DatadogLambdaProps, DatadogLambdaStrictProps, LambdaFunction } from "./interfaces";
 
 export const AWS_LAMBDA_EXEC_WRAPPER_KEY = "AWS_LAMBDA_EXEC_WRAPPER";
@@ -41,37 +42,6 @@ export const DD_LLMOBS_AGENTLESS_ENABLED = "DD_LLMOBS_AGENTLESS_ENABLED";
 const execSync = require("child_process").execSync;
 
 const URL = require("url").URL;
-
-// Tracks env vars written by this library per Function, so we can read them back without
-// touching aws-cdk-lib's private Function.environment field. aws-cdk-lib has no public
-// read accessor for env vars — the WeakMap is the library's own bookkeeping only.
-//
-// WeakMap (vs Map) so Function objects can be garbage-collected when a stack goes out of
-// scope (e.g. between test cases) without this module holding a lingering reference.
-//
-// Note: env vars set on a Function via func.addEnvironment() outside of this library
-// are not tracked here and will be overridden if the library sets the same key.
-// Configure DD_* vars via DatadogLambdaProps, or call func.addEnvironment() after
-// calling addLambdaFunctions().
-const ddEnvTracker: WeakMap<LambdaFunction, Map<string, string>> = new WeakMap();
-
-export function setTrackedEnv(lam: LambdaFunction, key: string, value: string): void {
-  let envMap = ddEnvTracker.get(lam);
-  if (!envMap) {
-    envMap = new Map();
-    ddEnvTracker.set(lam, envMap);
-  }
-  envMap.set(key, value);
-  lam.addEnvironment(key, value);
-}
-
-export function getTrackedEnv(lam: LambdaFunction, key: string): string | undefined {
-  return ddEnvTracker.get(lam)?.get(key);
-}
-
-export function hasTrackedEnv(lam: LambdaFunction, key: string): boolean {
-  return ddEnvTracker.get(lam)?.has(key) ?? false;
-}
 
 export function setGitEnvironmentVariables(
   lambdas: LambdaFunction[],
